Move code to google.ads directory
