+++ docs/autoapi_templates/index.rst
@@ -110,6 +110,10 @@ All operators are in the following packages:
 
   airflow/providers/ftp/sensors/index
 
+  airflow/providers/google/ads/operators/index
+
+  airflow/providers/google/ads/sensors/index
+
   airflow/providers/google/cloud/operators/index
 
   airflow/providers/google/cloud/sensors/index
@@ -246,6 +250,8 @@ All hooks are in the following packages:
   airflow/providers/elasticsearch/hooks/index
 
   airflow/providers/ftp/hooks/index
+
+  airflow/providers/google/ads/hooks/index
 
   airflow/providers/google/cloud/hooks/index
